Fixing display of item and skill buttons in char creation; moving showSkillPanel from UIElements to CharacterCreation since that's the only place it's used; fixing object panel to move max-height to object-list-container to adapt height to single item contents but limit height  when showing list
