+++ client/src/css/ui.css
@@ -214,6 +214,7 @@ h3 {
 	border-radius: 5px;
 	color: white;
 	background-color: #996633;
+	text-align: center;
 }
 
 .button-disabled {
@@ -449,6 +450,7 @@ h3 {
 
 .char-info-skills-skill-container {
 	display: flex;
+	gap: 10px;
 	padding-bottom: 20px;
 }
 
@@ -944,7 +946,6 @@ select {
 .object-info-panel, .skill-info-panel {
 	position: fixed;
 	width: 400px;
-	max-height: 300px;
 	padding: 10px;
 	background-color: #225555;
 	font-family: "Century Gothic", sans-serif;
@@ -965,30 +966,23 @@ select {
 	min-width: 300px;
 }
 
-.object-panel-contents {
+.object-list-container {
+	max-height: 300px;
+}
+
+.object-panel-contents, .skill-panel-contents {
 	display: flex;
 	justify-content: space-around;
 	margin-bottom: 20px;
-	flex-wrap: wrap;
 }
 
-.object-text-container {
+.object-text-container, .skill-text-container {
 	flex-basis: 70%;
 	display: flex;
 	flex-direction: column;
 	justify-content: space-between;
 	align-items: start;
 	gap: 5px;
-	max-height: 200px;
-	overflow-y: auto;
-	margin-bottom: 20px;
-}
-
-.skill-text-container {
-	display: flex;
-	flex-direction: column;
-	gap: 20px;
-	max-height: 200px;
 	overflow-y: auto;
 	margin-bottom: 20px;
 }
@@ -1007,7 +1001,8 @@ select {
 }
 
 .object-row-with-buttons {
-	display: flex;
+	display: grid;
+	grid-template-columns: 1fr 1fr 1fr 1fr;
 	width: 100%;
 	align-items: center;
 	gap: 10px;
